+++ storageold/defaults.go
@@ -1,4 +1,4 @@
-package storage
+package storageold
 
 import (
 	"github.com/dgraph-io/badger"