+++ support/globals/start.go
@@ -27,6 +27,7 @@ func Start() {
 	ShutdownTime = internalConfig.Section("buffers").Key("shutdown").MustInt(3)
 
 	SensorTimeout = internalConfig.Section("timeouts").Key("device").MustInt(5)
+	MaliciousTimeout = internalConfig.Section("timeouts").Key("malicious").MustInt(120)
 
 	CRCused = internalConfig.Section("sensors").Key("crc_enabled").MustBool(true)
 	fmt.Printf("*** WARNING: CRC usage is set to %v ***\n", CRCused)